Expanded Java testing support
Tests for Java methods now require that the parameter and return value types are specified along with the value. This allows for type conversion to take place whenever SnakeYAML does not support automatically creating an object from a custom class. This lays the groundwork for creating instances of custom classes in a test.